refactor(callers): migrate remaining grep_fast/glob_fast users to new homes
Phase F.2 of Phase 3 restructure plan.  Six other modules referenced
the deleted wrappers; all migrate to either glob_helpers (pure-Python
helpers) or direct nexus._rust_compat (Rust passthroughs).

  - backends/connectors/x/connector.py: glob_fast.glob_filter →
    glob_helpers.glob_filter (importlib trampoline preserved for the
    brick-tier-violation guard).
  - factory/_bricks.py: drops the try/except ImportError around
    `from .glob_fast import glob_fast` — glob_helpers is mandatory and
    nexus_kernel is mandatory, so the optional-import dance was dead.
    glob_match_fn now wires glob_helpers.glob_match directly.
  - factory/_distributed.py: docstring rename glob_fast → glob_helpers.
  - bricks/workflows/triggers.py: docstring rename "Rust glob_fast" →
    "glob_match function".  Public API unchanged: triggers still
    accept an injected GlobMatchFn.
  - server/api/v2/routers/async_files.py: split the import — glob_filter
    comes from glob_helpers, grep_files_mmap comes from
    nexus._rust_compat.
  - tests/benchmarks/test_search_benchmarks.py: every
    `from .grep_fast import grep_bulk / grep_files_mmap / RUST_AVAILABLE
    / MMAP_AVAILABLE` becomes a `from nexus._rust_compat import ...`
    plus a local `RUST_AVAILABLE = True` (benchmarks no longer probe
    availability — the kernel binary is required to build).
