+++ tests/benchmarks/test_search_benchmarks.py
@@ -126,13 +126,15 @@ class TestRustGrepBenchmarks:
 
     def test_rust_grep_available(self):
         """Check if Rust grep is available."""
-        from nexus.bricks.search.primitives.grep_fast import RUST_AVAILABLE
+        RUST_AVAILABLE = True
 
         print(f"\n[INFO] Rust grep available: {RUST_AVAILABLE}")
 
     def test_rust_grep_1k_lines(self, benchmark):
         """Benchmark Rust grep in 1K lines."""
-        from nexus.bricks.search.primitives.grep_fast import RUST_AVAILABLE, grep_bulk
+        from nexus._rust_compat import grep_bulk
+
+        RUST_AVAILABLE = True
 
         content = generate_log_content(1000)
         file_contents = {"/test.log": content}
@@ -161,7 +163,7 @@ class TestRustGrepBenchmarks:
     @pytest.mark.benchmark_ci
     def test_rust_grep_10k_lines(self, benchmark):
         """Benchmark Rust grep in 10K lines."""
-        from nexus.bricks.search.primitives.grep_fast import grep_bulk
+        from nexus._rust_compat import grep_bulk
 
         content = generate_log_content(10000)
         file_contents = {"/test.log": content}
@@ -182,7 +184,7 @@ class TestRustGrepBenchmarks:
 
     def test_rust_grep_multiple_files(self, benchmark):
         """Benchmark Rust grep across multiple files."""
-        from nexus.bricks.search.primitives.grep_fast import grep_bulk
+        from nexus._rust_compat import grep_bulk
 
         # Create 10 files with 1K lines each
         file_contents = {f"/file_{i}.log": generate_log_content(1000) for i in range(10)}
@@ -204,7 +206,7 @@ class TestRustGrepBenchmarks:
 
     def test_rust_grep_regex_pattern(self, benchmark):
         """Benchmark Rust grep with regex pattern."""
-        from nexus.bricks.search.primitives.grep_fast import grep_bulk
+        from nexus._rust_compat import grep_bulk
 
         content = generate_code_content(5000)
         file_contents = {"/code.py": content}
@@ -228,7 +230,7 @@ class TestRustGrepBenchmarks:
 
     def test_rust_grep_case_insensitive(self, benchmark):
         """Benchmark Rust grep with case-insensitive search."""
-        from nexus.bricks.search.primitives.grep_fast import grep_bulk
+        from nexus._rust_compat import grep_bulk
 
         content = generate_log_content(5000)
         file_contents = {"/test.log": content}
@@ -282,13 +284,15 @@ class TestRustMmapGrepBenchmarks:
 
     def test_mmap_grep_available(self):
         """Check if mmap grep is available."""
-        from nexus.bricks.search.primitives.grep_fast import MMAP_AVAILABLE
+        MMAP_AVAILABLE = True
 
         print(f"\n[INFO] Mmap grep available: {MMAP_AVAILABLE}")
 
     def test_mmap_grep_single_file(self, benchmark):
         """Benchmark mmap grep on a single file."""
-        from nexus.bricks.search.primitives.grep_fast import MMAP_AVAILABLE, grep_files_mmap
+        from nexus._rust_compat import grep_files_mmap
+
+        MMAP_AVAILABLE = True
 
         def search():
             result = grep_files_mmap("ERROR", [self.large_file], ignore_case=False)
@@ -315,7 +319,9 @@ class TestRustMmapGrepBenchmarks:
 
     def test_mmap_grep_multiple_files(self, benchmark):
         """Benchmark mmap grep across multiple files."""
-        from nexus.bricks.search.primitives.grep_fast import MMAP_AVAILABLE, grep_files_mmap
+        from nexus._rust_compat import grep_files_mmap
+
+        MMAP_AVAILABLE = True
 
         file_paths = list(self.test_files.keys())
 
@@ -345,11 +351,9 @@ class TestRustMmapGrepBenchmarks:
 
     def test_mmap_vs_bulk_grep_comparison(self, benchmark):
         """Compare mmap grep vs bulk grep (read + grep)."""
-        from nexus.bricks.search.primitives.grep_fast import (
-            MMAP_AVAILABLE,
-            grep_bulk,
-            grep_files_mmap,
-        )
+        from nexus._rust_compat import grep_bulk, grep_files_mmap
+
+        MMAP_AVAILABLE = True
 
         file_paths = list(self.test_files.keys())
 
@@ -379,7 +383,7 @@ class TestRustMmapGrepBenchmarks:
 
     def test_mmap_grep_case_insensitive(self, benchmark):
         """Benchmark mmap grep with case-insensitive search."""
-        from nexus.bricks.search.primitives.grep_fast import grep_files_mmap
+        from nexus._rust_compat import grep_files_mmap
 
         def search():
             result = grep_files_mmap("error", [self.large_file], ignore_case=True)
@@ -400,7 +404,7 @@ class TestRustMmapGrepBenchmarks:
 
     def test_mmap_grep_regex_pattern(self, benchmark):
         """Benchmark mmap grep with regex pattern."""
-        from nexus.bricks.search.primitives.grep_fast import grep_files_mmap
+        from nexus._rust_compat import grep_files_mmap
 
         # Create code files for regex testing
         code_content = generate_code_content(5000)
@@ -461,7 +465,9 @@ class TestGlobPatternBenchmarks:
 
     def test_rust_glob_simple(self, benchmark):
         """Benchmark Rust glob for simple patterns (if available)."""
-        from nexus.bricks.search.primitives.glob_fast import RUST_AVAILABLE, glob_match_bulk
+        from nexus._rust_compat import glob_match_bulk
+
+        RUST_AVAILABLE = True
 
         paths = [f"/dir/file_{i:04d}.txt" for i in range(1000)]
         paths += [f"/dir/file_{i:04d}.py" for i in range(1000)]
@@ -484,7 +490,7 @@ class TestGlobPatternBenchmarks:
 
     def test_rust_glob_multiple_patterns(self, benchmark):
         """Benchmark Rust glob with multiple patterns."""
-        from nexus.bricks.search.primitives.glob_fast import glob_match_bulk
+        from nexus._rust_compat import glob_match_bulk
 
         paths = [f"/dir/file_{i:04d}.txt" for i in range(500)]
         paths += [f"/dir/file_{i:04d}.py" for i in range(500)]
@@ -506,7 +512,7 @@ class TestGlobPatternBenchmarks:
 
     def test_rust_glob_recursive_pattern(self, benchmark):
         """Benchmark Rust glob with recursive pattern (**/*)."""
-        from nexus.bricks.search.primitives.glob_fast import glob_match_bulk
+        from nexus._rust_compat import glob_match_bulk
 
         # Generate paths with directory structure
         paths = []
